Removed unnecessary interface
